requirements.txt
streamlit>=1.22.0
numpy>=1.18.5
pillow>=7.1.2
requests>=2.24.0
python-dotenv>=0.15.0
pandas>=1.0.0